+++ crates/wormhole-node/benches/alloc_profile.rs
@@ -63,9 +63,14 @@ impl Transport {
 
     #[inline(always)]
     fn roundtrip(&mut self, payload: &[u8]) -> usize {
-        let enc_len = self.tx.write_message(payload, self.send.as_mut_slice()).unwrap();
+        let enc_len = self
+            .tx
+            .write_message(payload, self.send.as_mut_slice())
+            .unwrap();
         let cipher = unsafe { self.send.get_unchecked(..enc_len) };
-        self.rx.read_message(cipher, self.recv.as_mut_slice()).unwrap()
+        self.rx
+            .read_message(cipher, self.recv.as_mut_slice())
+            .unwrap()
     }
 }